Consolidate confirm controller
