+++ app/views/imports/_nav.html.erb
@@ -8,6 +8,6 @@
     <div class="h-px bg-alpha-black-200 w-12"></div>
     <%= render partial: "imports/nav_item", locals: { name: "Clean", path: import.persisted? ? clean_import_path(import) : nil, index: 3, is_complete: false, is_current: false } %>
     <div class="h-px bg-alpha-black-200 w-12"></div>
-    <%= render partial: "imports/nav_item", locals: { name: "Confirm", path: import.persisted? ? import_confirm_path(import) : nil, index: 4, is_complete: false, is_current: false } %>
+    <%= render partial: "imports/nav_item", locals: { name: "Confirm", path: import.persisted? ? confirm_import_path(import) : nil, index: 4, is_complete: false, is_current: false } %>
   </ul>
 </div>
