+++ app/views/imports/_preview_transaction.html.erb
