+++ app/views/imports/clean.html.erb
@@ -25,5 +25,5 @@
     <% end %>
   </table>
 
-  <%= link_to "Next", import_confirm_path(@import), class: "px-4 py-2 block w-full rounded-lg bg-gray-900 text-white text-sm font-medium", data: { turbo: false } %>
+  <%= link_to "Next", confirm_import_path(@import), class: "px-4 py-2 block w-full rounded-lg bg-gray-900 text-white text-sm font-medium", data: { turbo: false } %>
 </div>
