--- app/views/imports/confirms/show.html.erb
@@ -1,13 +1,0 @@
-<%= content_for :return_to_path, return_to_path(params, imports_path) %>
-
-<%= content_for :nav do %>
-  <%= render "imports/nav", import: @import %>
-<% end %>
-
-<div>
-  <h1 class="sr-only">Confirm import</h1>
-
-  <%= render partial: "imports/confirms/preview_transaction_group", collection: @import.preview_transactions.group_by(&:date), as: :transaction_group %>
-
-  <%= button_to "Import " + @import.preview_transactions.size.to_s + " transactions", import_confirm_path(@import), method: :patch, data: { turbo: false } %>
-</div>
